refactor : 패키지 이동 및 NumberExpression 함수 접근제거자 변경
