+++ src/main/java/calculator/expression/DivideExpression.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.expression;
 
 public class DivideExpression extends OperationExpression {
