+++ src/main/java/calculator/expression/Expression.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.expression;
 
 public interface Expression {
