+++ src/main/java/calculator/expression/OperationExpression.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.expression;
 
 public abstract class OperationExpression implements Expression {
