+++ src/test/java/calculator/expression/DivideExpressionTest.java
@@ -1,10 +1,19 @@
-package calculator;
+package calculator.expression;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
 public class DivideExpressionTest {
+
+    @Test
+    public void 나누기_테스트() {
+        Expression expression = new DivideExpression(new NumberExpression(5), new NumberExpression(3));
+
+        NumberExpression result = expression.calc();
+
+        assertThat(result).isEqualTo(new NumberExpression(5 / 3));
+    }
 
     @Test
     public void 다중_나누기_테스트() {
