+++ src/test/java/calculator/expression/MultiplyExpressionTest.java
@@ -1,10 +1,19 @@
-package calculator;
+package calculator.expression;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
 public class MultiplyExpressionTest {
+
+    @Test
+    public void 곱하기_테스트() {
+        Expression expression = new MultiplyExpression(new NumberExpression(5), new NumberExpression(5));
+
+        NumberExpression result = expression.calc();
+
+        assertThat(result).isEqualTo(new NumberExpression(5 * 5));
+    }
 
     @Test
     public void 연속된_곱하기_테스트() {
