+++ src/test/java/calculator/expression/PlusExpressionTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.expression;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,12 +7,21 @@ import org.junit.jupiter.api.Test;
 public class PlusExpressionTest {
 
     @Test
+    public void 더하기_테스트() {
+        PlusExpression expression = new PlusExpression(new NumberExpression(5), new NumberExpression(5));
+
+        NumberExpression result = expression.calc();
+
+        assertThat(result).isEqualTo(new NumberExpression(5 + 5));
+    }
+
+    @Test
     public void 연속된_더하기_테스트() {
         PlusExpression leftExpression = new PlusExpression(new NumberExpression(5), new NumberExpression(5));
         PlusExpression expression = new PlusExpression(leftExpression, new NumberExpression(5));
 
         NumberExpression result = expression.calc();
 
-        assertThat(result).isEqualTo(new NumberExpression(15));
+        assertThat(result).isEqualTo(new NumberExpression(5 + 5 + 5));
     }
 }
