+++ src/test/java/calculator/expression/SubtractExpressionTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.expression;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,13 +7,22 @@ import org.junit.jupiter.api.Test;
 public class SubtractExpressionTest {
 
     @Test
+    public void 빼기_테스트(){
+        Expression expression = new SubtractExpression(new NumberExpression(10),new NumberExpression(5));
+
+        NumberExpression calc = expression.calc();
+
+        assertThat(calc).isEqualTo(new NumberExpression(10 - 5));
+    }
+
+    @Test
     public void 연속된_빼기_테스트(){
         Expression leftExpression = new SubtractExpression(new NumberExpression(10),new NumberExpression(5));
         Expression expression = new SubtractExpression(leftExpression,new NumberExpression(5));
 
         NumberExpression calc = expression.calc();
 
-        assertThat(calc).isEqualTo(new NumberExpression(0));
+        assertThat(calc).isEqualTo(new NumberExpression(10 - 5 - 5));
     }
 
 }
